fix db store
